refactor internal packages
