+++ internal/geniustypes/pathspec.go
@@ -1,4 +1,4 @@
-package data
+package geniustypes
 
 import (
 	"fmt"
